Code optmized
