+++ src/components/map/mapbox/layers/pmtiles/__test__/PMTilesLayer.test.tsx
@@ -3,19 +3,6 @@ import { vi } from "vitest";
 import type { Map } from "mapbox-gl";
 import {
   buildPopupHtml,
-  periodNumberToDayjs,
-  parsePeriodInt,
-  parsePMTilesMetadata,
-  clampRangeToMetadata,
-  clampPeriodsToMetadata,
-  metadataRangeToDayjs,
-  sumSparseCountFromProperties,
-  sumCountsTreeInRange,
-  parseCountsTree,
-  formatPeriodInt,
-  coerceCountValue,
-  dayjsToPeriodInt,
-  daysInMonth,
   buildFeatureStateTotalExpression,
   buildFeatureStateTotalIsSetExpression,
   buildFeatureStateHasCountExpression,
@@ -27,8 +14,6 @@ import {
   updateFeatureStateTotals,
   scheduleDeferredWork,
   scheduleDebouncedWork,
-  buildCountFilterRange,
-  buildCountFilterRangeFromPeriods,
   createFeatureStateTotalsSession,
   featureStateSessionKey,
   resolvePmtilesFeatureId,
@@ -51,6 +36,21 @@ import {
   DAYS_KEY,
   TOTAL_KEY,
   densityStopValue,
+  periodNumberToDayjs,
+  parsePeriodInt,
+  parsePMTilesMetadata,
+  clampRangeToMetadata,
+  clampPeriodsToMetadata,
+  metadataRangeToDayjs,
+  sumSparseCountFromProperties,
+  sumCountsTreeInRange,
+  parseCountsTree,
+  formatPeriodInt,
+  coerceCountValue,
+  dayjsToPeriodInt,
+  daysInMonth,
+  buildCountFilterRange,
+  buildCountFilterRangeFromPeriods,
   PMTilesMetadataRange,
 } from "../Common";
