+++ src/pages/detail-page/features/MapPanel.tsx
@@ -64,10 +64,11 @@ import {
   DatasetType,
   OGCCollection,
 } from "@/app/store/OGCCollectionDefinitions";
-import PMTilesHexLayer, {
+import PMTilesHexLayer from "@/components/map/mapbox/layers/pmtiles/PMTilesLayer";
+import {
   metadataRangeToDayjs,
-  PMTilesMetadata,
-} from "@/components/map/mapbox/layers/pmtiles/PMTilesLayer";
+  type PMTilesMetadata,
+} from "@/components/map/mapbox/layers/pmtiles/Common";
 
 const mapContainerId = "map-detail-container-id";
